% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agespectrum.R
\name{agespectrum}
\alias{agespectrum}
\alias{agespectrum.ArAr}
\alias{agespectrum.default}
\title{Plot a (\eqn{^{40}}Ar/\eqn{^{39}}Ar) release spectrum}
\usage{
agespectrum(x, ...)

\method{agespectrum}{default}(x, alpha = 0.05, plateau = TRUE,
  plateau.col = rgb(0, 1, 0, 0.5), non.plateau.col = rgb(0, 1, 1, 0.5),
  sigdig = 2, line.col = "red", lwd = 2, ...)

\method{agespectrum}{ArAr}(x, alpha = 0.05, plateau = TRUE,
  plateau.col = rgb(0, 1, 0, 0.5), non.plateau.col = rgb(0, 1, 1, 0.5),
  sigdig = 2, dcu = TRUE, line.col = "red", lwd = 2, ...)
}
\arguments{
\item{x}{a three column matrix whose first column gives the amount
    of \eqn{^{39}}Ar in each aliquot, and whose second and third
    columns give the age and its uncertainty.

OR

an object of class \code{ArAr} with \code{format=2}}

\item{...}{optional parameters to the generic plot function}

\item{alpha}{the confidence limits of the error bars/boxes.}

\item{plateau}{Boolean flag indicating whether a plateau age should
be calculated if \code{plateau=TRUE}, the function will compute
the weighted mean of the largest succession of steps that yield
values passing the Chi-square test for age homogeneity.}

\item{plateau.col}{the fill colour of the rectangles used to mark
the steps belonging to the age plateau.}

\item{non.plateau.col}{if \code{plateau=TRUE}, the steps that do
NOT belong to the plateau are given a different colour.}

\item{sigdig}{the number of significant digits of the numerical
values reported in the title of the graphical output
(only used if \code{plateau=FALSE}).}

\item{line.col}{colour of the isochron line}

\item{lwd}{line width}

\item{dcu}{propagate the decay constant uncertainties?}
}
\description{
Produces a plot of boxes whose widths correspond to the cumulative
amount of \eqn{^{39}}Ar (or any other volume proxy), and whose
heights express the analytical uncertainties.
}
\examples{
data(examples)
agespectrum(examples$ArAr)
}

