% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{as.UPb}
\alias{as.UPb}
\title{Create object of class 'UPb'}
\usage{
as.UPb(x, format = 1)
}
\arguments{
\item{x}{matrix with U-Pb data}

\item{format}{formatting option

1: 'Pb207Pb206','errPb207Pb206', 'Pb206U238','errPb206U238', 'Pb207U235','errPb207U235'}
}
\description{
Create a 'UPb' object from a matrix with U-Pb data
}
\examples{
fname <- system.file("UPb.csv",package="IsoplotR")
dat <- read.csv(fname,row.names=1,header=TRUE)
concordia.plot(as.UPb(dat))
}

