% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPb.R
\name{concordant}
\alias{concordant}
\title{Test whether the U-Pb compositions are concordant}
\usage{
concordant(x, cutoff.76 = 1100, cutoff.disc = c(-15, 5))
}
\arguments{
\item{x}{an object of class \code{UPb}}

\item{cutoff.76}{the age (in Ma) below which the concordance
between the \eqn{^{206}}Pb/\eqn{^{238}}U and
\eqn{^{207}}Pb/\eqn{^{238}}U ages are tested, and above which
the concordance between the \eqn{^{206}}Pb/\eqn{^{238}}U and
\eqn{^{207}}Pb/\eqn{^{206}}Pb ages is tested.}

\item{cutoff.disc}{two element vector with the minimum (negative)
and maximum (positive) percentage discordance allowed between
the \eqn{^{207}}Pb/\eqn{^{235}}U and
\eqn{^{206}}Pb/\eqn{^{238}}U age (if
\eqn{^{206}}Pb/\eqn{^{238}}U < \code{cutoff.76}) or between the
\eqn{^{206}}Pb/\eqn{^{238}}U and \eqn{^{207}}Pb/\eqn{^{206}}Pb
age (if \eqn{^{206}}Pb/\eqn{^{238}}U > \code{cutoff.76}).}
}
\value{
A vector of boolean flags marking the concordant
    aliquots.
}
\description{
Assess whether U-Pb \eqn{{}^{206}}Pb/\eqn{{}^{238}}U- and
\eqn{{}^{207}}Pb/\eqn{{}^{206}}Pb-ages, or the
\eqn{{}^{206}}Pb/\eqn{{}^{238}}U- and
\eqn{{}^{207}}Pb/\eqn{{}^{235}}U ages are concordant.
}
\details{
Returns \code{TRUE} if the
    \eqn{{}^{206}}Pb/\eqn{{}^{238}}U and
    \eqn{{}^{207}}Pb/\eqn{{}^{235}}U-ages agree within the limits
    specified in \code{cutoff.disc} (if
    t[\eqn{{}^{206}}Pb/\eqn{{}^{238}}U] < \code{cutoff.76}, OR if
    the \eqn{{}^{206}}Pb/\eqn{{}^{238}}U and
    \eqn{{}^{207}}Pb/\eqn{{}^{206}}Pb-ages agree within the limits
    specified in \code{cutoff.disc} (if
    t[\eqn{{}^{206}}Pb/\eqn{{}^{238}}U] > \code{cutoff.76}.
    Returns \code{FALSE} otherwise.
}
\examples{
data(examples)
conc <- concordant(examples$UPb,cutoff.disc=c(-1,1))
radialplot(examples$UPb,omit=which(!conc))
}
