% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordia.R
\name{concordia.age}
\alias{concordia.age}
\alias{concordia.age.UPb}
\alias{concordia.age.default}
\title{Calculate U-Pb concordia ages}
\usage{
concordia.age(x, ...)

\method{concordia.age}{default}(x, wetherill = TRUE, dcu = TRUE, ...)

\method{concordia.age}{UPb}(x, wetherill = TRUE, dcu = TRUE, i = NA, ...)
}
\arguments{
\item{x}{either an object of class \code{UPb}, or a list containing
two items, \code{x} and \code{cov}, corresponding to a U-Pb
composition and its covariance matrix, respectively.}

\item{...}{optional arguments}

\item{wetherill}{boolean flag to indicate whether the data should
be evaluated in Wetherill (\code{TRUE}) or Tera-Wasserburg
(\code{FALSE}) space}

\item{dcu}{propagate the decay constant uncertainties?}

\item{i}{(optional) scalar index of a particular aliquot. If omitted, the
concordia age corresponding to the weighted mean composition is calculated}
}
\value{
a list with the following items:

\code{x}: a named vector with the (weighted mean) U-Pb composition

\code{cov}: the covariance matrix of the (mean) U-Pb composition

\code{age}: the concordia age (in Ma)

\code{age.err}: the standard error of the concordia age

And, additionally (if \code{x} has class \code{UPb} and
\code{!is.na(i)}):

\code{mswd}: a list with two items (\code{equivalence} and
\code{concordance}) containing the MSWD (Mean of the Squared
Weighted Deviates, a.k.a the reduced Chi-squared statistic outside
of geochronology) of isotopic equivalence and age concordance,
respectively.

\code{p.value}: a list with two items (\code{equivalence} and
\code{concordance}) containing the p-value of the Chi-square test
for isotopic equivalence and age concordance, respectively.
}
\description{
Evaluates the equivalence of multiple
(\eqn{^{206}}Pb/\eqn{^{238}}U-\eqn{^{207}}Pb/\eqn{^{235}}U or
\eqn{^{207}}Pb/\eqn{^{206}}Pb-\eqn{^{206}}Pb/\eqn{^{238}}U)
compositions, computes the weighted mean isotopic composition and
the corresponding concordia age using the method of maximum
likelihood, computes the mswd of equivalence and concordance and
their respective Chi-squared p-values.
}
\examples{
data(examples)
fit <- concordia.age(examples$UPb)
print(paste('age = ',fit$age,'+/-',fit$age.err,'Ma, MSWD = ',fit$mswd))
}

