% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{examples}
\alias{examples}
\title{Example datasets for testing \code{IsoplotR}}
\description{
U-Pb and detrital zircon datasets
}
\details{
\code{examples} is a list with two items

\code{UPb}: an object of class \code{'UPb'} containing a high
precision U-Pb dataset packaged with Ken Ludwig's \code{Isoplot}
program.

\code{DZ}: an object of class \code{'detrital'} containing a
detrital zircon U-Pb dataset from Namibia.
}
\examples{
data(examples)
concordia(examples$UPb)
dev.new()
kde(examples$DZ)
}
\author{
Ken Ludwig and Pieter Vermeesch
}
\references{
Ludwig, K. R. User's manual for Isoplot 3.00: a geochronological
    toolkit for Microsoft Excel. No. 4. Kenneth R. Ludwig, 2003.

Vermeesch, Pieter, and Eduardo
Garzanti. "Making geological sense of 'Big Data' in sedimentary provenance analysis."
Chemical Geology 409 (2015): 20-27.
}

