% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{get.covmat}
\alias{get.covmat}
\title{Get the covariance matrix of a sample}
\usage{
get.covmat(x, i)
}
\arguments{
\item{x}{an object of class \code{UPb}}

\item{i}{the index of the sample of interest}
}
\value{
a covariance matrix of size [3x3]
}
\description{
Returns the covariance matrix of the i\eqn{^{th}} sample
}
\examples{
data(UPb)
get.covmat(UPb,2)
}

