% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPb.R
\name{get.covmat.UPb}
\alias{get.covmat.UPb}
\title{Get the covariance matrix of a U/Pb samples}
\usage{
get.covmat.UPb(X, i)
}
\arguments{
\item{X}{an object of class 'UPb'}

\item{i}{the index of the sample of interest}
}
\value{
a 3x3 covariance matrix
}
\description{
Returns the covariance matrix of the U/Pb and Pb/Pb ratios of the ith sample
}
\examples{
data(UPb)
get.covmat.UPb(UPb,2)
}

