% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPb.R
\name{get.Pb206U238}
\alias{get.Pb206U238}
\title{Calculate a Pb206/U238 ratio}
\usage{
get.Pb206U238(age)
}
\arguments{
\item{age}{a U/Pb age [in Ma]}
}
\value{
a Pb206/U238 ratio
}
\description{
Returns the expected Pb206/U238 ratio for a given age
}
\examples{
get.Pb206U238(4567)
}

