% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPb.R
\name{get.Pb206U238age}
\alias{get.Pb206U238age}
\title{Calculate a Pb206/U238 age}
\usage{
get.Pb206U238age(Pb206U238)
}
\arguments{
\item{Pb206U238}{a scalar}
}
\value{
the Pb206/U238 age [in Ma]
}
\description{
Returns the U/Pb age for a given Pb206/U238 ratio
}
\examples{
get.Pb206U238age(1.03)
}

