% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPb.R
\name{get.Pb207Pb206}
\alias{get.Pb207Pb206}
\title{Calculate a Pb207/Pb206 ratio}
\usage{
get.Pb207Pb206(age)
}
\arguments{
\item{age}{a Pb/Pb age [in Ma]}
}
\value{
a Pb207/Pb206 ratio
}
\description{
Returns the expected Pb207/Pb206 ratio for a given age
}
\examples{
get.Pb207Pb206(4567)
}

