% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPb.R
\name{get.Pb207Pb206age}
\alias{get.Pb207Pb206age}
\title{Calculate a Pb207/Pb206 age}
\usage{
get.Pb207Pb206age(Pb207Pb206)
}
\arguments{
\item{Pb207Pb206}{the Pb207/Pb206 ratio}
}
\value{
the Pb207/Pb206 age [in Ma]
}
\description{
Numerically solves the Pb207/Pb206 age equation
}
\examples{
get.Pb207Pb206age(0.625)
}

