% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPb.R
\name{get.Pb207U235}
\alias{get.Pb207U235}
\title{Calculate a Pb207/U235 ratio}
\usage{
get.Pb207U235(age)
}
\arguments{
\item{age}{a U/Pb age [in Ma]}
}
\value{
a Pb207/U235 ratio
}
\description{
Returns the expected Pb207/U235 ratio for a given age
}
\examples{
get.Pb207U235(4567)
}

