% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPb.R
\name{get.Pb207U235age}
\alias{get.Pb207U235age}
\title{Calculate a Pb207/U235 age}
\usage{
get.Pb207U235age(Pb207U235)
}
\arguments{
\item{Pb207U235}{a scalar}
}
\value{
the Pb207/U235 age [in Ma]
}
\description{
Returns the U/Pb age for a given Pb207/U235 ratio
}
\examples{
get.Pb207U235age(88.8)
}

