% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{get.ratios}
\alias{get.ratios}
\title{Calculate the isotopic ratio for a given age}
\usage{
get.ratios(age, method = "U-Pb")
}
\arguments{
\item{age}{the geological age [Ma]}

\item{method}{currently only \code{'U-Pb'}, other chronometers will
be added later}
}
\value{
a two element list containing:

\code{x}: a vector with predicted isotopic ratios

\code{cov}: the covariance matrix of the predicted isotopic ratios
taking into account decay constant uncertainties
}
\description{
Predict the daughter/parent ratio for a given U-Pb age
}
\examples{
get.ratios(4567,'U-Pb')
}

