% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{I.A}
\alias{I.A}
\title{Isotope abundance}
\usage{
I.A(nuclide, x = NULL, e = NULL)
}
\arguments{
\item{nuclide}{one of either \code{'U'}, \code{'U238'},
\code{'U235'}, or \code{'Th232'}}

\item{x}{new value for the isotope abundance}

\item{e}{new value for the standard error of the abundance}
}
\value{
if x == e == NULL, returns a two element list containing:

\code{x}: a number or a vector of numbers between 0 (absent) and 1 (dominant)

and

\code{e}: the standard error or covariance matrix of x

or, if \code{nuclide} = \code{'U'}:

\code{cov}: the covariance matrix of all naturally occurring
isotopes
}
\description{
Gets or sets the natural abundance of isotopes
}
\examples{
print(I.A('U238')$x)
# use the 238U/235U ratio of Steiger and Jaeger (1977)
U238U235(138.88,0)
print(I.A('U238')$x)
}

