% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoplotR.R
\docType{package}
\name{IsoplotR}
\alias{IsoplotR}
\alias{_PACKAGE}
\alias{IsoplotR-package}
\alias{IsoplotR-package}
\title{library(IsoplotR)}
\description{
A list of documented functions may be viewed by typing
\code{help(package='IsoplotR')}.  Detailed instructions are
provided at \url{http://isoplotr.london-geochron.com}. A manuscript
with the theoretical background is in review for publication in
\emph{Geoscience Frontiers}.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://isoplotr.london-geochron.com}
}

}
\author{
\strong{Maintainer}: Pieter Vermeesch \email{p.vermeesch@ucl.ac.uk}

}
