% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde.plot}
\alias{kde.plot}
\title{Plot (a) kernel density estimate(s)}
\usage{
kde.plot(x, from = NA, to = NA, bw = NA, adaptive = TRUE, log = FALSE,
  n = 512, samebandwidth = TRUE, normalise = FALSE, binwidth = NA, ...)
}
\arguments{
\item{x}{an object of class \code{UPb} or \code{detritals}}

\item{from}{minimum age of the time axis. If NULL, this is set
automatically}

\item{to}{maximum age of the time axis. If NULL, this is set
automatically}

\item{bw}{the bandwidth of the KDE. If NULL, bw will be calculated
automatically using \code{botev()}}

\item{adaptive}{boolean flag controlling if the adaptive KDE
modifier of Abramson (1982) is used}

\item{log}{transform the ages to a log scale if TRUE}

\item{n}{horizontal resolution of the density estimate}

\item{samebandwidth}{boolean flag indicating whether the same
bandwidth should be used for all samples. If samebandwidth =
TRUE and bw = NULL, then the function will use the median
bandwidth of all the samples. This option is only used if
\code{x} is of class \code{detritals}.}

\item{normalise}{boolean flag indicating whether or not the KDEs
should all integrate to the same value. This option is only
used if \code{x} is of class \code{detritals}.}

\item{binwidth}{scalar width of the histogram bins, in Myr if
\code{x$log==FALSE}, or as a fractional value if
\code{x$log==TRUE}. Sturges' Rule is used if
\code{binwidth==NA}}

\item{...}{optional arguments to be passed on to \code{plot.KDEs}
(if \code{x} is of class \code{detritals} or \code{plot.KDE}
otherwise}
}
\description{
Plots geochronological datsets as kernel density estimates using a
combination of the Botev (2010) bandwidth selector and the Abramson
(1982) adaptive kernel bandwidth modifier.
}
\examples{
data(examples)
kde.plot(examples$DZ[['N2']])
}
\seealso{
kde kdes plot.KDE plot.KDEs
}

