% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kdeplot}
\alias{kdeplot}
\title{Generate and plot (a) kernel density estimate(s)}
\usage{
kdeplot(x, from = NA, to = NA, bw = NA, adaptive = TRUE, log = FALSE,
  n = 512, pch = NA, xlab = "age [Ma]", ylab = "", kde.col = rgb(1, 0,
  1, 0.6), hist.col = rgb(0, 1, 0, 0.2), show.hist = TRUE, bty = "n",
  binwidth = NA, ncol = NA, ...)
}
\arguments{
\item{x}{an object of class \code{UPb} or \code{detritals}}

\item{from}{minimum age of the time axis. If NULL, this is set
automatically}

\item{to}{maximum age of the time axis. If NULL, this is set
automatically}

\item{bw}{the bandwidth of the KDE. If NULL, bw will be calculated
automatically using \code{botev()}}

\item{adaptive}{boolean flag controlling if the adaptive KDE
modifier of Abramson (1982) is used}

\item{log}{transform the ages to a log scale if TRUE}

\item{n}{horizontal resolution of the density estimate}

\item{pch}{the symbol used to show the samples. May be a vector.
Set \code{pch = NA} to turn them off.}

\item{xlab}{the label of the x-axis}

\item{ylab}{the label of the y-axis}

\item{kde.col}{the fill colour of the KDE specified as a four
element vector of r, g, b, alpha values}

\item{hist.col}{the fill colour of the histogram specified as a
four element vector of r, g, b, alpha values}

\item{show.hist}{boolean flag indicating whether a histogram should
be added to the KDE}

\item{bty}{change to \code{"o"}, \code{"l"}, \code{"7"},
\code{"c"}, \code{"u"}, or \code{"]"} if you want to draw a box
around the plot}

\item{binwidth}{scalar width of the histogram bins, in Myr if
\code{x$log==FALSE}, or as a fractional value if
\code{x$log==TRUE}. Sturges' Rule is used if
\code{binwidth==NA}}

\item{ncol}{scalar value indicating the number of columns over
which the KDEs should be divided. This option is only used if
\code{x} is of class \code{detritals}.}

\item{...}{optional arguments to be passed on to \code{kde(x,...)}}
}
\description{
Plots geochronological datsets as kernel density estimates using a
combination of the Botev (2010) bandwidth selector and the Abramson
(1982) adaptive kernel bandwidth modifier.
}
\examples{
data(examples)
kdeplot(examples$DZ[['N2']])
}

