% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kdes}
\alias{kdes}
\title{Create a list of KDEs}
\usage{
kdes(x, from = NA, to = NA, bw = NA, samebandwidth = TRUE,
  adaptive = TRUE, normalise = FALSE, log = FALSE, n = 512, ...)
}
\arguments{
\item{x}{a named list of vectors containing ordinal data}

\item{from}{minimum limit of the x-axis.}

\item{to}{maximum limit of the x-axis.}

\item{bw}{the bandwidth of the kernel density estimates. If bw =
NA, the bandwidth will be set automatically using \code{botev()}}

\item{samebandwidth}{boolean flag indicating whether the same
bandwidth should be used for all samples. If samebandwidth = TRUE
and bw = NULL, then the function will use the median bandwidth of
all the samples.}

\item{adaptive}{boolean flag switching on the adaptive bandwidth
modifier of Abramson (1982)}

\item{normalise}{boolean flag indicating whether or not the KDEs
should all integrate to the same value.}

\item{log}{boolean flag indicating whether the data should by
plotted on a logarithmic scale.}

\item{n}{horizontal resolution of the density estimates}

\item{...}{optional parameters to be passed on to \code{density}}
}
\value{
an object of class \code{KDEs}, i.e. a list containing the
following items:

kdes: a named list with objects of class \code{KDE}

from: the beginning of the common time scale

to: the end of the common time scale

themax: the maximum probability density of all the KDEs

xlabel: the x-axis label to be used by \code{plot.KDEs}
}
\description{
Convert a list of numerical vectors into a list of objects of class
\code{KDE}
}
\examples{
data(examples)
KDES <- kdes(examples$DZ,from=0,to=3000)
plot(KDES)
}
\seealso{
kde
}

