% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{lambda.err}
\alias{lambda.err}
\title{Decay constant uncertainties}
\usage{
lambda.err(nuclide)
}
\arguments{
\item{nuclide}{the nuclide name}
}
\value{
the analytical uncertainty of the decay constant  [in Ma-1]
}
\description{
Returns the standard errors of radioactive decay constants
}
\examples{
lambda.err('U238')
}

