% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{lambda}
\alias{lambda}
\title{Decay constants}
\usage{
lambda(nuclide)
}
\arguments{
\item{nuclide}{the nuclide name}
}
\value{
the decay constant  [in Ma-1]
}
\description{
Returns the decay constants of radioactive istopes
}
\examples{
lambda('U238')
}

