% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{length}
\alias{length}
\alias{length.UPb}
\alias{length.PbPb}
\alias{length.ArAr}
\alias{length.PD}
\alias{length.ThPb}
\alias{length.KCa}
\alias{length.ThU}
\alias{length.UThHe}
\alias{length.KDE}
\alias{length.KDEs}
\alias{length.fissiontracks}
\title{Count the number of aliquots in a sample}
\usage{
length(x)

length.UPb(x)

length.PbPb(x)

length.ArAr(x)

length.PD(x)

length.ThPb(x)

length.KCa(x)

length.ThU(x)

length.UThHe(x)

length.KDE(x)

length.KDEs(x)

length.fissiontracks(x)
}
\arguments{
\item{x}{an object of class \code{UPb}, \code{PbPb}, \code{ArAr},
\code{ThPb}, \code{KCa}, \code{PD}, \code{ThU},
\code{detritals}, \code{UThHe} or \code{fissiontracks}.}
}
\value{
a scalar
}
\description{
Count the number of rows (aliquots) in a dataset
    returned by \code{read.data} or, for datasets of class
    \code{detritals}, the number of columns (samples).
}
\examples{
data(examples)
length(examples$UPb)
}
