% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ludwig.R
\name{ludwig}
\alias{ludwig}
\title{Linear regression of X,Y,Z-variables with correlated errors, taking
into account decay constant uncertainties}
\usage{
ludwig(x, covmat)
}
\arguments{
\item{x}{a \eqn{3n}-element vector \eqn{[X Y Z]}, where \eqn{X},
\eqn{Y} and \eqn{Z} are three \eqn{n}-element vectors of
(isotopic ratio) values.}

\item{covmat}{a \eqn{[3n x 3n]}-element covariance matrix of
\code{x}}
}
\description{
Implements the maximum likelihood algorithm of Ludwig (1998)
}
\references{
Ludwig, K.R., 1998. On the treatment of concordant uranium-lead
ages. Geochimica et Cosmochimica Acta, 62(4), pp.665-676.
}
