% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPb.R
\name{names.UPb}
\alias{names.UPb}
\title{Get the names of the samples contained in a U/Pb dataset}
\usage{
\method{names}{UPb}(x)
}
\arguments{
\item{x}{an object of class 'UPb'}
}
\value{
an object of class 'UPb'
}
\description{
Returns the rownames of a U/Pb dataset
}
\examples{
data(UPb)
names(UPb)
}

