% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogls.R
\name{ogls}
\alias{ogls}
\alias{ogls.default}
\alias{ogls.other}
\title{Omnivariant Generalized Least-Squares Regression}
\usage{
ogls(x, ...)

\method{ogls}{default}(x, model = 1, ...)

\method{ogls}{other}(x, model = 1, ...)
}
\arguments{
\item{x}{either a \code{n x (n+1)} matrix obtained by prepending a
vector of alternating \code{X,Y}-values to its covariance
matrix OR an object of class \code{other} with
\code{x$format=6}.}

\item{...}{optional arguments}

\item{model}{construct the isochron using either:

\code{1}: Ordinary OGLS regression

\code{2}: Total least squares regression

\code{3}: OGLS regression with an overdispersion term}
}
\value{
a list of the slope and intercept of the best fit line as
    well as their standard errors and covariance.
}
\description{
Linear regression with inter-sample error correlations.
}
\examples{
fn <- system.file('UW137.csv',package='IsoplotR')
UW137 <- read.data(fn,method='other',format=6)
fit <- ogls(UW137)
}
