% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonPb.R
\name{Pb0corr}
\alias{Pb0corr}
\title{Common Pb correction}
\usage{
Pb0corr(x, option = 1, omit = NULL)
}
\arguments{
\item{x}{an object of class \code{UPb}}

\item{option}{one of either

\enumerate{
   \item Stacey-Kramers
   \item isochron
   \item nominal
}}

\item{omit}{vector with indices of aliquots that should be omitted
from the isochron regression (only used if \code{option==2})}
}
\value{
Returns a list in which \code{x.raw} contains the original data and
\code{x} the common Pb-corrected compositions. All other items in
the list are inherited from the input data.
}
\description{
Applies a common-Pb correction to a U-Pb dataset using either the
Stacey-Kramers mantle evolution model, isochron regression, or any
nominal inital Pb isotope composition.
}
\details{
...
}
\examples{
data(examples)
UPb <- Pb0corr(examples$UPb,option=1)
concordia(UPb)
# produces identical results as:
dev.new()
concordia(examples$UPb,common.Pb=1)
}
