% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfit.R
\name{peakfit}
\alias{peakfit}
\alias{peakfit.ArAr}
\alias{peakfit.UPb}
\alias{peakfit.UThHe}
\alias{peakfit.default}
\alias{peakfit.fissiontracks}
\title{Finite mixture modelling of geochronological datasets}
\usage{
peakfit(x, ...)

\method{peakfit}{default}(x, k = 1, ...)

\method{peakfit}{fissiontracks}(x, k = 1, exterr = TRUE, ...)

\method{peakfit}{UPb}(x, k = 1, type = 4, cutoff.76 = 1100,
  cutoff.disc = c(-15, 5), exterr = TRUE, ...)

\method{peakfit}{ArAr}(x, k = 1, exterr = TRUE, ...)

\method{peakfit}{UThHe}(x, k = 1, ...)
}
\arguments{
\item{x}{either a \code{[2 x n]} matrix with measurements and their
standard errors, or an object of class \code{fissiontracks}}

\item{k}{the number of discrete age components to be
sought. Setting this parameter to \code{'auto'} automatically
selects the optimal number of components (up to a maximum of 5)
using the Bayes Information Criterion (BIC).}

\item{exterr}{propagate the external sources of uncertainty into
the component age errors?}
}
\description{
Implements the discrete mixture modelling algorithms of Galbraith
and Green (1993) and applies them to fission track and other
geochronological datasets.
}
\examples{
data(examples)
peakfit(examples$FT1,k=2)
}
\references{
Galbraith, R.F. and Laslett, G.M., 1993. Statistical models for
mixed fission track ages. Nuclear tracks and radiation
measurements, 21(4), pp.459-470.
}

