% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{plot.KDE}
\alias{plot.KDE}
\title{Plot a kernel density estimate}
\usage{
\method{plot}{KDE}(x, pch = "|", xlab = "age [Ma]", ylab = "",
  kde.col = rgb(1, 0, 1, 0.6), show.hist = TRUE, hist.col = rgb(0, 1, 0,
  0.2), binwidth = NA, bty = "n", ...)
}
\arguments{
\item{x}{an object of class \code{KDE}}

\item{pch}{the symbol used to show the samples. May be a vector.
Set \code{pch = NA} to turn them off.}

\item{xlab}{the label of the x-axis}

\item{ylab}{the label of the y-axis}

\item{kde.col}{the fill colour of the KDE specified as a four
element vector of r, g, b, alpha values}

\item{show.hist}{boolean flag indicating whether a histogram should
be added to the KDE}

\item{hist.col}{the fill colour of the histogram specified as a
four element vector of r, g, b, alpha values}

\item{binwidth}{scalar width of the histogram bins, in Myr if
\code{x$log==FALSE}, or as a fractional value if
\code{x$log==TRUE}. Sturges' Rule is used if \code{binwidth==NA}}

\item{bty}{change to \code{"o"}, \code{"l"}, \code{"7"},
\code{"c"}, \code{"u"}, or \code{"]"} if you want to draw a box
around the plot}

\item{...}{optional parameters to be passed on to the graphics
object}
}
\description{
Plots an object of class \code{KDE}
}
\examples{
data(examples)
dens <- kde(examples$DZ[['N1']],from=0,to=3000)
plot(dens)
}
\seealso{
KDE
}

