% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{settings}
\alias{settings}
\title{Load settings to and from json}
\usage{
settings(fname = NULL)
}
\arguments{
\item{fname}{the path of a \code{.json} file}
}
\value{
if \code{fname==NULL}, returns a \code{.json} string
}
\description{
Get and set preferred values for decay constants and isotopic
abundances from and to a \code{.json} file format
}
\examples{
json <- system.file("defaults.json",package="IsoplotR")
settings(json)
print(settings())
}

