% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{subset.UPb}
\alias{subset.UPb}
\alias{subset.PbPb}
\alias{subset.ArAr}
\alias{subset.ThPb}
\alias{subset.KCa}
\alias{subset.PD}
\alias{subset.ThU}
\alias{subset.detritals}
\alias{subset.UThHe}
\alias{subset.fissiontracks}
\title{Subset geochronological datasets}
\usage{
\method{subset}{UPb}(x, ...)

\method{subset}{PbPb}(x, ...)

\method{subset}{ArAr}(x, ...)

\method{subset}{ThPb}(x, ...)

\method{subset}{KCa}(x, ...)

\method{subset}{PD}(x, ...)

\method{subset}{ThU}(x, ...)

\method{subset}{detritals}(x, select, ...)

\method{subset}{UThHe}(x, ...)

\method{subset}{fissiontracks}(x, ...)
}
\arguments{
\item{x}{an object of class \code{UPb}, \code{PbPb}, \code{ArAr},
\code{ThPb}, \code{KCa}, \code{PD}, \code{ThU},
\code{detritals}, \code{UThHe} or \code{fissiontracks}.}

\item{...}{arguments to be passed onto the generic
    \code{\link[base]{subset}} function. Specifically:

\code{subset}: logical expression indicating elements or rows to
keep: missing values are taken as false.

\code{select}: expression, indicating columns to select from a data frame.

Note: if \code{x} has class \code{detritals}, then only the
argument \code{select} is accepted. In this case the function will
also accept the indices of the samples to retain.}

\item{select}{a vector with sample names or numbers.}
}
\value{
an object with the same class as \code{x}.
}
\description{
Retain selected rows or columns from a larger dataset
    produced by \code{read.data}.
}
\examples{
data(examples)
tt <- age(examples$UPb)
old <- (tt[,'t.conc'] > 230)
UPb <- subset(examples$UPb,subset=old)
}
