% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{U238U235}
\alias{U238U235}
\title{238U/235 ratio}
\usage{
U238U235(x = NULL, e = NULL)
}
\arguments{
\item{x}{new value for 238U/235U ratio}

\item{e}{new value for its standard error}
}
\value{
if x == e == NULL, returns a two-item list containing:

\code{x}: the 238U/235U ratio

\code{e}: the standard error of the 238U/235U ratio
}
\description{
Gets or sets the natural 238U/235 ratio. The default value of
137.818 is taken from Hiess et al. (2012)
}
\examples{
print(U238U235()$x)
# use the 238U/235U ratio of Steiger and Jaeger (1977)
U238U235(138.88,0)
print(U238U235()$x)
}

