% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{yorkfit}
\alias{yorkfit}
\title{Linear regression of X,Y-variables with correlated errors}
\usage{
yorkfit(X, Y, sX, sY, rXY)
}
\arguments{
\item{X}{vector of measurements}

\item{Y}{vector of measurements}

\item{sX}{standard errors of \code{X}}

\item{sY}{standard errors of \code{Y}}

\item{rXY}{correlation coefficients between X and Y}
}
\value{
a five element list containing

\code{a}: the intercept of the straight line fit

\code{b}: the slope of the fit

\code{sa}: the standard error of the intercept

\code{sb}: the standard error of the slope
}
\description{
Implements the unified regression algorithm of York et al. (2004)
which, although based on least squares, yields results that are
consistent with maximum likelihood estimates of Ludwig and
Titterington (1994)
}
\examples{
   X <- c(1.550,12.395,20.445,20.435,20.610,24.900,
          28.530,50.540,51.595,86.51,106.40,157.35)
   Y <- c(.7268,.7849,.8200,.8156,.8160,.8322,
          .8642,.9584,.9617,1.135,1.230,1.490)
   n <- length(X)
   sX <- X*0.01
   sY <- Y*0.005
   rXY <- rep(0.8,n)
   fit <- yorkfit(X,Y,sX,sY,rXY)
   covmat <- matrix(0,2,2)
   plot(range(X),fit$a+fit$b*range(X),type='l',ylim=range(Y))
   for (i in 1:n){
       covmat[1,1] <- sX[i]^2
       covmat[2,2] <- sY[i]^2
       covmat[1,2] <- rXY[i]*sX[i]*sY[i]
       covmat[2,1] <- covmat[1,2]
       ell <- ellipse(X[i],Y[i],covmat,alpha=0.05)
       polygon(ell)
   }
}
\references{
Ludwig, K. R., and D. M. Titterington. "Calculation of 230ThU
isochrons, ages, and errors." Geochimica et Cosmochimica Acta
58.22 (1994): 5031-5042.

York, Derek, et al. "Unified equations for the slope,
intercept, and standard errors of the best straight line."
American Journal of Physics 72.3 (2004): 367-375.
}

