% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\alias{age.UPb}
\alias{age.default}
\title{Calculate isotopic ages}
\usage{
age(x, ...)

\method{age}{default}(x, ...)

\method{age}{UPb}(x, concordia = 0, i = NA, ...)
}
\arguments{
\item{x}{an object of class \code{UPb}}

\item{...}{optional arguments}

\item{concordia}{one of either

0: consider each U-Pb analysis separately

1: calculate a concordia age from all U-Pb analyses together

2: fit a discordia line through all the U-Pb analyses}

\item{i}{(optional) index of a particular aliquot}
}
\value{
if \code{x} has class \code{UPb}, \code{concordia}=0,
    returns a table with the following columns: 't.75',
    'err[t.75]','t.68','err[t.68]', 't.76','err[t.76]',
    't.conc','err[t.conc]', containing the 207Pb/235U-age
    and standard error, the 206Pb/238U-age and standard error,
    the 207Pb/206Pb-age and standard error, and the concordia
    age and standard error, respectively.

if \code{x} has class \code{UPb} and \code{concordia=1}, returns
the output of \link{concordia.age}

if \code{x} has class \code{UPb} and \code{concordia=2}, returns
the output of \link{discordia.age}
}
\description{
Calculates U-Pb and Ar-Ar ages
}
\examples{
data(examples)
age(examples$UPb)
}

