% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPb.R
\name{age}
\alias{age}
\alias{age.UPb}
\alias{age.default}
\title{Calculate isotopic ages}
\usage{
age(x, ...)

\method{age}{default}(x, ...)

\method{age}{UPb}(x, concordia = 0, ...)
}
\arguments{
\item{x}{an object of class \code{UPb}}

\item{...}{optional arguments}

\item{concordia}{one of either

0: consider each U-Pb analysis separately

1: calculate a concordia age from all U-Pb analyses together

2: fit a discordia line through all the U-Pb analyses}
}
\value{
if \code{x} has class \code{UPb} and \code{concordia} = 0,
    a table with the following columns:
}
\description{
Calculates U-Pb ages
}

