% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\alias{age.UPb}
\alias{age.default}
\title{Calculate isotopic ages}
\usage{
age(x, ...)

\method{age}{default}(x, ...)

\method{age}{UPb}(x, concordia = 0, i = NA, ...)
}
\arguments{
\item{x}{an object of class \code{UPb}}

\item{...}{optional arguments}

\item{concordia}{one of either

0: consider each U-Pb analysis separately

1: calculate a concordia age from all U-Pb analyses together

2: fit a discordia line through all the U-Pb analyses}

\item{i}{(optional) index of a particular aliquot}
}
\value{
if \code{x} has class \code{UPb}, \code{concordia}=0, and
    \code{i=NA}, returns a table with the following columns:
    '7/5-age', 'err[7/5-age]', '6/8-age', 'err[6/8-age]',
    '7/6-age', 'err[7/6-age]', 'concordia-age', and
    'err[concordia-age]'.

if \code{x} has class \code{UPb}, \code{concordia}=0, and
    \code{i!=NA}, returns a list with two items (\code{x} and
    \code{cov}) containing the '7/5-age', '6/8-age', '7/6-age', and
    'concordia-age', and their covariance matrix, respectively.

if \code{x} has class \code{UPb} and \code{concordia=1}, returns
the output of \link{concordia.age}

if \code{x} has class \code{UPb} and \code{concordia=2}, returns
the output of \link{discordia.age}
}
\description{
Calculates U-Pb and Ar-Ar ages
}
\examples{
data(examples)
age(examples$UPb)
}

