% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/york.R
\name{data2york}
\alias{data2york}
\title{Prepare geochronology data for York regression}
\usage{
data2york(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}

\item{format}{one of

1. \code{X}, \code{s[X]}, \code{Y}, \code{s[Y]}, \code{rho}

or

2. \code{X/Z}, \code{s[X/Z]}, \code{Y/Z}, \code{s[Y/Z]},
\code{X/Y}, \code{s[X/Y]}}

\item{wetherill}{TODO}

\item{inverse}{TODO}

\item{exterr}{TODO}

\item{common}{TODO}
}
\value{
TODO
}
\description{
Takes geochronology data as input and produces a
five-column table as output, which can be used
for York regression.
}
