% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{evolution}
\alias{evolution}
\title{Th-U evolution diagram}
\usage{
evolution(x, xlim = NA, ylim = NA, alpha = 0.05, transform = FALSE,
  detrital = FALSE, show.numbers = FALSE, levels = NA,
  ellipse.col = c("#00FF0080", "#FF000080"), line.col = "darksalmon",
  isochron = FALSE, model = 1, exterr = TRUE, sigdig = 2, ...)
}
\arguments{
\item{x}{an object of class \code{ThU}}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{alpha}{probability cutoff for the error ellipses and
confidence intervals}

\item{transform}{if \code{TRUE}, plots \eqn{^{234}}U/\eqn{^{238}}U
vs. Th-U age.}

\item{detrital}{apply a detrital Th correction by projecting the
compositions along an isochron?}

\item{show.numbers}{label the error ellipses with the grain
numbers?}

\item{levels}{a vector with additional values to be displayed as
different background colours within the error ellipses.}

\item{ellipse.col}{a vector of two background colours for the error
ellipses. If \code{levels=NA}, then only the first colour will
be used. If \code{levels} is a vector of numbers, then
\code{ellipse.col} is used to construct a colour ramp.}

\item{line.col}{colour of the age grid}

\item{isochron}{fit a 3D isochron to the data?}

\item{model}{if \code{isochron=TRUE}, choose one of three
    regression models:

\code{1}: maximum likelihood regression, using either the modified
error weighted least squares algorithm of York et al. (2004) for
2-dimensional data, or the Maximum Likelihood formulation of Ludwig
and Titterington (1994) for 3-dimensional data. These algorithms
take into account the analytical uncertainties and error
correlations, under the assumption that the scatter between the
data points is solely caused by the analytical uncertainty. If the
assumption is correct, then the MSWD value should be approximately
equal to one. There are three strategies to deal with the case
where MSWD>1. The first of these is to assume that the analytical
uncertainties have been underestimated by a factor
\eqn{\sqrt{MSWD}}. Alternative approaches are described below.

\code{2}: ordinary least squares regression: a second way to deal
with over- or underdispersed datasets is to simply ignore the
analytical uncertainties.

\code{3}: maximum likelihood regression with overdispersion:
instead of attributing any overdispersion (MSWD > 1) to
underestimated analytical uncertainties (model 1), one can also
attribute it to the presence of geological uncertainty, which
manifests itself as an added (co)variance term.}

\item{exterr}{propagate the decay constant uncertainty in the
isochron age?}

\item{sigdig}{number of significant digits for the isochron age}

\item{...}{optional arguments to the generic \code{plot} function}
}
\description{
Plots Th-U data on a
\eqn{^{234}}U/\eqn{^{238}}U-\eqn{^{230}}Th/\eqn{^{238}}U evolution
diagram, a \eqn{^{234}}U/\eqn{^{238}}U-age diagram, or
(if \eqn{^{234}}U/\eqn{^{238}}U is assumed to be in secular equilibrium),
a \eqn{^{230}}Th/\eqn{^{232}}Th-\eqn{^{238}}U/\eqn{^{232}}Th diagram,
calculates isochron ages.
}
\examples{
data(examples)
evolution(examples$ThU)
}
\references{
Ludwig, K.R. and Titterington, D.M., 1994. Calculation
    of \eqn{^{230}}Th/U isochrons, ages, and errors. Geochimica et
    Cosmochimica Acta, 58(22), pp.5031-5042.

Ludwig, K.R., 2003. Mathematical-statistical treatment of data and
    errors for \eqn{^{230}}Th/U geochronology. Reviews in Mineralogy and
    Geochemistry, 52(1), pp.631-656.
}
