% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{evolution}
\alias{evolution}
\title{Th-U evolution diagram}
\usage{
evolution(x, xlim = NA, ylim = NA, alpha = 0.05, transform = FALSE,
  show.numbers = FALSE, ellipse.col = rgb(0, 1, 0, 0.5),
  line.col = "darksalmon", isochron = FALSE, exterr = TRUE, sigdig = 2,
  ...)
}
\arguments{
\item{x}{an object of class \code{ThU}}

\item{xlim}{x-axis limits}

\item{ylim}{y-axis limits}

\item{alpha}{confidence cutoff for the error ellipses}

\item{transform}{if \code{TRUE}, plots
\eqn{^{234}}U/\eqn{^{238}}U vs. Th-U age.}

\item{show.numbers}{logical flag (\code{TRUE} to show grain
numbers)}

\item{ellipse.col}{background colour of the error ellipses}

\item{line.col}{colour of the age grid}

\item{isochron}{fit a 3D isochron to the data?}

\item{exterr}{show decay constant uncertainty?}

\item{sigdig}{number of significant digits for the isochron age}

\item{...}{optional arguments to the generic \code{plot} function}
}
\description{
Plots Th-U data on a
\eqn{^{234}}U/\eqn{^{238}}U-\eqn{^{230}}Th/\eqn{^{238}}U evolution
diagram or a \eqn{^{234}}U/\eqn{^{238}}U-age diagram, calculates
isochron ages.
}
\examples{
data(examples)
evolution(examples$ThU)
}
\references{
Ludwig, K.R., 2003. Mathematical-statistical treatment
    of data and errors for 230 Th/U geochronology. Reviews in
    Mineralogy and Geochemistry, 52(1), pp.631-656.
}
