% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{I.A}
\alias{I.A}
\title{Isotope abundance}
\usage{
I.A(nuclide)
}
\arguments{
\item{nuclide}{the nuclide name}
}
\value{
a number between 0 (absent) and 1 (dominant)
}
\description{
Returns the natural abundance of isotopes
}
\examples{
I.A('U238')
}

