% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{I.A}
\alias{I.A}
\title{Isotope abundance}
\usage{
I.A(nuclide)
}
\arguments{
\item{nuclide}{one of either \code{'U'}, \code{'U238'},
\code{'U235'}, or \code{'Th232'}}
}
\value{
a two element list containing:

\code{x}: a number or a vector of numbers between 0 (absent) and 1 (dominant)

and

\code{e}: the standard error or covariance matrix of x

or, if \code{nuclide} = \code{'U'}:

\code{cov}: the covariance matrix of all naturally occurring
isotopes
}
\description{
Returns the natural abundance of isotopes
}
\examples{
I.A('U238')
}

