% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochron.R
\name{isochron}
\alias{isochron}
\alias{isochron.ArAr}
\alias{isochron.default}
\title{Calculate and plot isochrons}
\usage{
isochron(x, ...)

\method{isochron}{default}(x, xlim = NA, ylim = NA, alpha = 0.05,
  show.numbers = FALSE, ellipse.col = rgb(0, 1, 0, 0.5),
  line.col = "grey", lwd = 2, ...)

\method{isochron}{ArAr}(x, xlim = NA, ylim = NA, alpha = 0.05,
  show.numbers = FALSE, ellipse.col = rgb(0, 1, 0, 0.5), inverse = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{x}{EITHER a list with the following vectors:

\code{X:} the x-variable

\code{Y:} the y-variable

\code{sX:} the standard error of X

\code{sY:} the standard error of Y

\code{rXY:} the correlation coefficient of X and Y

OR an object of class \code{ArAr}}

\item{...}{optional arguments}

\item{xlim}{2-element vector with the plot limits of the x-axis}

\item{ylim}{2-element vector with the plot limits of the y-axis}

\item{alpha}{confidence cutoff for the error ellipses}

\item{show.numbers}{boolean flag (TRUE to show grain numbers)}

\item{ellipse.col}{background colour of the error ellipses}

\item{line.col}{colour of the isochron line}

\item{lwd}{line width}

\item{inverse}{if \code{TRUE}, plots \eqn{^{36}}Ar/\eqn{^{40}}Ar
vs. \eqn{^{39}}Ar/\eqn{^{40}}Ar. If \code{FALSE}, plots
\eqn{^{40}}Ar/\eqn{^{36}}Ar vs. \eqn{^{39}}Ar/\eqn{^{36}}Ar.}

\item{plot}{if \code{FALSE}, suppresses the graphical output}
}
\value{
if \code{plot=FALSE}, returns a list with the following items:

\code{a:} the intercept of the straight line fit and its standard error

\code{b:} the slope of the fit and its standard error

\code{y0:} the atmospheric \eqn{^{40}}Ar/\eqn{^{36}}Ar ratio and its standard error

\code{age:} the \eqn{^{40}}Ar/\eqn{^{39}}Ar age and its standard error
}
\description{
Plots cogenetic \eqn{^{40}}Ar/\eqn{^{39}}Ar data as X-Y
scatterplots, fits an isochron curve through them using the
\code{yorkfit} function, and computes the corresponding isochron
age, including decay constant uncertainties.
}
\examples{
data(examples)
isochron(examples$ArAr)
}

