% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{IsoplotR-internal}
\alias{IsoplotR-internal}
\alias{Abramson}
\alias{ArAr.age}
\alias{ArAr.age.ratios}
\alias{ArAr.inverse.ratios}
\alias{ArAr.normal.ratios}
\alias{BIC_fit}
\alias{EDM.age}
\alias{ICP.age}
\alias{KS.diss}
\alias{LL.concordia.age}
\alias{LL.concordia.comp}
\alias{LL.concordia.comp.default}
\alias{LL.concordia.comp.terawasserburg}
\alias{LL.concordia.comp.wetherill}
\alias{LL.concordia.intersection.york}
\alias{LL.lud.UPb}
\alias{LL.norm}
\alias{LL.weightedmean}
\alias{LuHf.age}
\alias{PD.age}
\alias{Pb.correction.with.204}
\alias{Pb.correction.without.204}
\alias{PbPb.age}
\alias{PbPb.inverse.ratios}
\alias{PbPb.normal.ratios}
\alias{RbSr.age}
\alias{ReOs.age}
\alias{S.tit}
\alias{SS.UThHe.uv}
\alias{SS.UThHe.uvw}
\alias{SmNd.age}
\alias{ThU.age}
\alias{ThU.convert}
\alias{ThU.gr}
\alias{ThU.misfit}
\alias{U4U8vsTh0U8}
\alias{U4U8vst}
\alias{UPb.age}
\alias{UThHe.age}
\alias{UThHe2uv}
\alias{UThHe2uv.covmat}
\alias{UThHe2uvw}
\alias{UThHe2uvw.covmat}
\alias{add.exterr}
\alias{age.PD}
\alias{age2radial}
\alias{age_to_Pb206U238_ratio}
\alias{age_to_Pb207Pb206_ratio}
\alias{age_to_Pb207U235_ratio}
\alias{age_to_U238Pb206_ratio}
\alias{age_to_concordia_ratios}
\alias{age_to_terawasserburg_ratios}
\alias{age_to_wetherill_ratios}
\alias{ages2peaks}
\alias{alpha.beta.gamma}
\alias{as.ArAr}
\alias{as.LuHf}
\alias{as.PD}
\alias{as.PbPb}
\alias{as.RbSr}
\alias{as.ReOs}
\alias{as.SmNd}
\alias{as.ThU}
\alias{as.UPb}
\alias{as.UThHe}
\alias{as.detritals}
\alias{as.fissiontracks}
\alias{as.other}
\alias{att}
\alias{binomial.mixtures}
\alias{botev}
\alias{common.Pb.correction}
\alias{common.Pb.isochron}
\alias{common.Pb.nominal}
\alias{common.Pb.stacey.kramers}
\alias{commonbandwidth}
\alias{concordia.age}
\alias{concordia.age.UPb}
\alias{concordia.age.default}
\alias{concordia.comp}
\alias{concordia.intersection}
\alias{concordia.intersection.ludwig}
\alias{concordia.intersection.york}
\alias{concordia.line}
\alias{concordia.title}
\alias{cor2cov2}
\alias{cor2cov3}
\alias{dD76dR}
\alias{dD76dl5}
\alias{dD76dl8}
\alias{dD76dt}
\alias{data2evolution}
\alias{data2ludwig}
\alias{data2ludwig.UPb}
\alias{data2ludwig.default}
\alias{data2ludwig_with_decay_err}
\alias{data2ludwig_without_decay_err}
\alias{data2rxry}
\alias{data2tit}
\alias{data2tit.ThU}
\alias{data2tit.default}
\alias{data2york}
\alias{data2york.ArAr}
\alias{data2york.PD}
\alias{data2york.PbPb}
\alias{data2york.UPb}
\alias{data2york.default}
\alias{dct1d}
\alias{discordia.plot}
\alias{discordia.title}
\alias{diss}
\alias{doSm}
\alias{eq.6.9}
\alias{errorprop}
\alias{etchfact}
\alias{evolution.lines}
\alias{evolution.title}
\alias{f147Sm}
\alias{filter.UPb.ages}
\alias{fisher.lud}
\alias{fisher.lud.UPb}
\alias{fisher.lud.default}
\alias{fisher.tit}
\alias{fisher_lud_with_decay_err}
\alias{fisher_lud_without_decay_err}
\alias{fissiontrack.age}
\alias{fixedpoint}
\alias{fromJSON}
\alias{geomean.Sm}
\alias{get.ArAr.age}
\alias{get.ArAr.ratio}
\alias{get.EDM.age}
\alias{get.He}
\alias{get.ICP.age}
\alias{get.LuHf.age}
\alias{get.LuHf.ratio}
\alias{get.PD.age}
\alias{get.PD.ratio}
\alias{get.Pb206U238.age}
\alias{get.Pb206U238.age.UPb}
\alias{get.Pb206U238.age.default}
\alias{get.Pb206U238.age.terawasserburg}
\alias{get.Pb206U238.age.wetherill}
\alias{get.Pb206U238.ratios}
\alias{get.Pb207Pb206.age}
\alias{get.Pb207Pb206.age.UPb}
\alias{get.Pb207Pb206.age.default}
\alias{get.Pb207Pb206.age.terawasserburg}
\alias{get.Pb207Pb206.age.wetherill}
\alias{get.Pb207Pb206.ratios}
\alias{get.Pb207U235.age}
\alias{get.Pb207U235.age.UPb}
\alias{get.Pb207U235.age.default}
\alias{get.Pb207U235.age.wetherill}
\alias{get.Pb207U235.ratios}
\alias{get.RbSr.age}
\alias{get.RbSr.ratio}
\alias{get.ReOs.age}
\alias{get.ReOs.ratio}
\alias{get.SmNd.age}
\alias{get.SmNd.ratio}
\alias{get.Th230U238}
\alias{get.ThU.age}
\alias{get.U234U238}
\alias{get.U238Pb206.ratios}
\alias{get.UThHe.age}
\alias{get.UsU}
\alias{get.absolute.zeta}
\alias{get.concordia.SS}
\alias{get.concordia.limits}
\alias{get.cor.68.76}
\alias{get.cor.75.68}
\alias{get.cor.div}
\alias{get.cor.mult}
\alias{get.cov.46.76}
\alias{get.cov.46.86}
\alias{get.cov.68.48}
\alias{get.cov.75.48}
\alias{get.cov.75.68}
\alias{get.cov.76.86}
\alias{get.cov.div}
\alias{get.cov.mult}
\alias{get.helioplot.contours}
\alias{get.helioplot.tticks}
\alias{get.limits}
\alias{get.logratio.contours}
\alias{get.logratio.tticks}
\alias{get.logratioplot.limits}
\alias{get.max.z}
\alias{get.min.z}
\alias{get.minage.L}
\alias{get.offset}
\alias{get.peakfit.covmat}
\alias{get.props.err}
\alias{get.radial.tticks}
\alias{get.tticks}
\alias{get.weightedmean}
\alias{get.york.mswd}
\alias{get.york.xy}
\alias{get.z0}
\alias{getG}
\alias{getfact}
\alias{getkde}
\alias{getkde.default}
\alias{getkde.detritals}
\alias{getmM}
\alias{gr.tit}
\alias{hasClass}
\alias{helioplot_title}
\alias{iatt}
\alias{imass}
\alias{initial.concordia.age}
\alias{initial.concordia.age.default}
\alias{initial.concordia.age.terawasserburg}
\alias{initial.concordia.age.wetherill}
\alias{intersection.misfit.ludwig}
\alias{intersection.misfit.york}
\alias{iratio}
\alias{isochron_PD}
\alias{isochrontitle}
\alias{k1}
\alias{lambda}
\alias{length.ArAr}
\alias{length.LuHf}
\alias{length.PD}
\alias{length.PbPb}
\alias{length.RbSr}
\alias{length.ReOs}
\alias{length.SmNd}
\alias{length.ThU}
\alias{length.UPb}
\alias{length.UThHe}
\alias{length.fissiontracks}
\alias{matrix2covlist}
\alias{min_age_model}
\alias{min_age_to_legend}
\alias{mindens}
\alias{mswd.concordia}
\alias{mswd.lud}
\alias{mswd.tit}
\alias{newJSONParser}
\alias{normal.mixtures}
\alias{peakfit_helper}
\alias{peaks2legend}
\alias{pilotdensity}
\alias{plateau}
\alias{plateau.title}
\alias{plot.KDE}
\alias{plot.KDEs}
\alias{plot.MDS}
\alias{plot_central_ellipse}
\alias{plot_helioplot_contours}
\alias{plot_helioplot_ellipses}
\alias{plot_helioplot_frame}
\alias{plot_logratio_contours}
\alias{plot_logratio_ellipses}
\alias{plot_logratio_frame}
\alias{plot_points}
\alias{plot_radial_axes}
\alias{plot_radial_lines}
\alias{plot_weightedmean}
\alias{plotlines}
\alias{ppm2ratios}
\alias{ppm2ratios.LuHf}
\alias{ppm2ratios.RbSr}
\alias{ppm2ratios.ReOs}
\alias{ppm2ratios.SmNd}
\alias{ppm2ratios.default}
\alias{project.concordia}
\alias{radial.plot}
\alias{radial.scale}
\alias{radial.title}
\alias{radialplot_helper}
\alias{regression}
\alias{renormalise}
\alias{roundit}
\alias{scatterplot}
\alias{shiny2matrix}
\alias{t2z}
\alias{tera.wasserburg}
\alias{theta2age}
\alias{tit2york}
\alias{toJSON}
\alias{tracklength}
\alias{uv2HeUTh}
\alias{uv2UThHe}
\alias{uvw2UThHe}
\alias{weightedmean_helper}
\alias{wetherill}
\alias{wtdmean.title}
\alias{x2zs}
\alias{x2zs.default}
\alias{x2zs.fissiontracks}
\alias{xyz2xy}
\alias{z2rxy}
\title{Internal code objects}
\description{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written ).
}
\keyword{internal}
