% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoplotR.R
\docType{package}
\name{IsoplotR}
\alias{IsoplotR}
\alias{_PACKAGE}
\alias{IsoplotR-package}
\title{library(IsoplotR)}
\description{
A list of documented functions may be viewed by typing
\code{help(package='IsoplotR')}.  Detailed instructions are
provided at \url{http://isoplotr.london-geochron.com}. Further
details about the theoretical background are provided by Vermeesch
(2018).
}
\references{
Vermeesch, P., 2018, IsoplotR: a free and open toolbox
    for geochronology. Geoscience Frontiers, 9, 1479-1493, doi:
    10.1016/j.gsf.2018.04.001.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://isoplotr.london-geochron.com}
}

}
\author{
\strong{Maintainer}: Pieter Vermeesch \email{p.vermeesch@ucl.ac.uk}

}
