% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde}
\alias{kde}
\alias{kde.default}
\alias{kde.UPb}
\alias{kde.detritals}
\alias{kde.PbPb}
\alias{kde.ArAr}
\alias{kde.ThU}
\alias{kde.ReOs}
\alias{kde.SmNd}
\alias{kde.RbSr}
\alias{kde.LuHf}
\alias{kde.UThHe}
\alias{kde.fissiontracks}
\title{Create (a) kernel density estimate(s)}
\usage{
kde(x, ...)

\method{kde}{default}(x, from = NA, to = NA, bw = NA, adaptive = TRUE,
  log = FALSE, n = 512, plot = TRUE, pch = NA, xlab = "age [Ma]",
  ylab = "", kde.col = rgb(1, 0, 1, 0.6), hist.col = rgb(0, 1, 0, 0.2),
  show.hist = TRUE, bty = "n", binwidth = NA, ncol = NA, ...)

\method{kde}{UPb}(x, from = NA, to = NA, bw = NA, adaptive = TRUE,
  log = FALSE, n = 512, plot = TRUE, pch = NA, xlab = "age [Ma]",
  ylab = "", kde.col = rgb(1, 0, 1, 0.6), hist.col = rgb(0, 1, 0, 0.2),
  show.hist = TRUE, bty = "n", binwidth = NA, ncol = NA, type = 4,
  cutoff.76 = 1100, cutoff.disc = c(-15, 5), ...)

\method{kde}{detritals}(x, from = NA, to = NA, bw = NA, adaptive = TRUE,
  log = FALSE, n = 512, plot = TRUE, pch = NA, xlab = "age [Ma]",
  ylab = "", kde.col = rgb(1, 0, 1, 0.6), hist.col = rgb(0, 1, 0, 0.2),
  show.hist = TRUE, bty = "n", binwidth = NA, ncol = NA,
  samebandwidth = TRUE, normalise = TRUE, ...)

\method{kde}{PbPb}(x, from = NA, to = NA, bw = NA, adaptive = TRUE,
  log = FALSE, n = 512, plot = TRUE, pch = NA, xlab = "age [Ma]",
  ylab = "", kde.col = rgb(1, 0, 1, 0.6), hist.col = rgb(0, 1, 0, 0.2),
  show.hist = TRUE, bty = "n", binwidth = NA, ncol = NA, i2i = FALSE,
  ...)

\method{kde}{ArAr}(x, from = NA, to = NA, bw = NA, adaptive = TRUE,
  log = FALSE, n = 512, plot = TRUE, pch = NA, xlab = "age [Ma]",
  ylab = "", kde.col = rgb(1, 0, 1, 0.6), hist.col = rgb(0, 1, 0, 0.2),
  show.hist = TRUE, bty = "n", binwidth = NA, ncol = NA, i2i = FALSE,
  ...)

\method{kde}{ThU}(x, from = NA, to = NA, bw = NA, adaptive = TRUE,
  log = FALSE, n = 512, plot = TRUE, pch = NA, xlab = "age [ka]",
  ylab = "", kde.col = rgb(1, 0, 1, 0.6), hist.col = rgb(0, 1, 0, 0.2),
  show.hist = TRUE, bty = "n", binwidth = NA, ncol = NA, i2i = FALSE,
  ...)

\method{kde}{ReOs}(x, from = NA, to = NA, bw = NA, adaptive = TRUE,
  log = FALSE, n = 512, plot = TRUE, pch = NA, xlab = "age [Ma]",
  ylab = "", kde.col = rgb(1, 0, 1, 0.6), hist.col = rgb(0, 1, 0, 0.2),
  show.hist = TRUE, bty = "n", binwidth = NA, ncol = NA, i2i = TRUE,
  ...)

\method{kde}{SmNd}(x, from = NA, to = NA, bw = NA, adaptive = TRUE,
  log = FALSE, n = 512, plot = TRUE, pch = NA, xlab = "age [Ma]",
  ylab = "", kde.col = rgb(1, 0, 1, 0.6), hist.col = rgb(0, 1, 0, 0.2),
  show.hist = TRUE, bty = "n", binwidth = NA, ncol = NA, i2i = TRUE,
  ...)

\method{kde}{RbSr}(x, from = NA, to = NA, bw = NA, adaptive = TRUE,
  log = FALSE, n = 512, plot = TRUE, pch = NA, xlab = "age [Ma]",
  ylab = "", kde.col = rgb(1, 0, 1, 0.6), hist.col = rgb(0, 1, 0, 0.2),
  show.hist = TRUE, bty = "n", binwidth = NA, ncol = NA, i2i = TRUE,
  ...)

\method{kde}{LuHf}(x, from = NA, to = NA, bw = NA, adaptive = TRUE,
  log = FALSE, n = 512, plot = TRUE, pch = NA, xlab = "age [Ma]",
  ylab = "", kde.col = rgb(1, 0, 1, 0.6), hist.col = rgb(0, 1, 0, 0.2),
  show.hist = TRUE, bty = "n", binwidth = NA, ncol = NA, i2i = TRUE,
  ...)

\method{kde}{UThHe}(x, from = NA, to = NA, bw = NA, adaptive = TRUE,
  log = FALSE, n = 512, plot = TRUE, pch = NA, xlab = "age [Ma]",
  ylab = "", kde.col = rgb(1, 0, 1, 0.6), hist.col = rgb(0, 1, 0, 0.2),
  show.hist = TRUE, bty = "n", binwidth = NA, ncol = NA, ...)

\method{kde}{fissiontracks}(x, from = NA, to = NA, bw = NA,
  adaptive = TRUE, log = FALSE, n = 512, plot = TRUE, pch = NA,
  xlab = "age [Ma]", ylab = "", kde.col = rgb(1, 0, 1, 0.6),
  hist.col = rgb(0, 1, 0, 0.2), show.hist = TRUE, bty = "n",
  binwidth = NA, ncol = NA, ...)
}
\arguments{
\item{x}{a vector of numbers OR an object of class \code{UPb},
\code{PbPb}, \code{ArAr}, \code{ReOs}, \code{SmNd},
\code{RbSr}, \code{UThHe}, \code{fissiontracks}, \code{ThU} or
\code{detrital}}

\item{...}{optional arguments to be passed on to \code{density}}

\item{from}{minimum age of the time axis. If \code{NULL}, this is set
automatically}

\item{to}{maximum age of the time axis. If \code{NULL}, this is set
automatically}

\item{bw}{the bandwidth of the KDE. If \code{NULL}, \code{bw} will be calculated
automatically using \code{botev()}}

\item{adaptive}{logical flag controlling if the adaptive KDE
modifier of Abramson (1982) is used}

\item{log}{transform the ages to a log scale if \code{TRUE}}

\item{n}{horizontal resolution of the density estimate}

\item{plot}{show the KDE as a plot}

\item{pch}{the symbol used to show the samples. May be a vector.
Set \code{pch=NA} to turn them off.}

\item{xlab}{the x-axis label}

\item{ylab}{the y-axis label}

\item{kde.col}{the fill colour of the KDE specified as a four
element vector of \code{r, g, b, alpha} values}

\item{hist.col}{the fill colour of the histogram specified as a
four element vector of r, g, b, alpha values}

\item{show.hist}{logical flag indicating whether a histogram should
be added to the KDE}

\item{bty}{change to \code{"o"}, \code{"l"}, \code{"7"},
\code{"c"}, \code{"u"}, or \code{"]"} if you want to draw a box
around the plot}

\item{binwidth}{scalar width of the histogram bins, in Myr if
\code{x$log = FALSE}, or as a fractional value if
\code{x$log = TRUE}. Sturges' Rule is used if
\code{binwidth = NA}}

\item{ncol}{scalar value indicating the number of columns over
which the KDEs should be divided. This option is only used if
\code{x} has class \code{detritals}.}

\item{type}{scalar indicating whether to plot the
\eqn{^{207}}Pb/\eqn{^{235}}U age (\code{type}=1), the
\eqn{^{206}}Pb/\eqn{^{238}}U age (\code{type}=2), the
\eqn{^{207}}Pb/\eqn{^{206}}Pb age (\code{type}=3), the
\eqn{^{207}}Pb/\eqn{^{206}}Pb-\eqn{^{206}}Pb/\eqn{^{238}}U age
(\code{type}=4), or the (Wetherill) concordia age
(\code{type}=5)}

\item{cutoff.76}{the age (in Ma) below which the
\eqn{^{206}}Pb/\eqn{^{238}}U and above which the
\eqn{^{207}}Pb/\eqn{^{206}}Pb age is used. This parameter is
only used if \code{type=4}.}

\item{cutoff.disc}{two element vector with the maximum and minimum
percentage discordance allowed between the
\eqn{^{207}}Pb/\eqn{^{235}}U and \eqn{^{206}}Pb/\eqn{^{238}}U
age (if \eqn{^{206}}Pb/\eqn{^{238}}U < \code{cutoff.76}) or
between the \eqn{^{206}}Pb/\eqn{^{238}}U and
\eqn{^{207}}Pb/\eqn{^{206}}Pb age (if
\eqn{^{206}}Pb/\eqn{^{238}}U > \code{cutoff.76}).  Set
\code{cutoff.disc=NA} if you do not want to use this filter.}

\item{samebandwidth}{logical flag indicating whether the same
bandwidth should be used for all samples. If
\code{samebandwidth = TRUE} and \code{bw = NULL}, then the
function will use the median bandwidth of all the samples.}

\item{normalise}{logical flag indicating whether or not the KDEs
should all integrate to the same value.}

\item{i2i}{`isochron to intercept': calculates the initial
(aka `inherited', `excess', or `common') \eqn{^{40}}Ar/\eqn{^{36}}Ar,
\eqn{^{207}}Pb/\eqn{^{204}}Pb, \eqn{^{87}}Sr/\eqn{^{86}}Sr,
\eqn{^{143}}Nd/\eqn{^{144}}Nd, \eqn{^{187}}Os/\eqn{^{188}}Os or
\eqn{^{176}}Hf/\eqn{^{177}}Hf ratio from an isochron
fit. Setting \code{i2i} to \code{FALSE} uses the default values
stored in \code{settings('iratio',...)}  or zero (for the Pb-Pb
method). When applied to data of class \code{ThU}, setting
\code{i2i} to \code{TRUE} applies a detrital Th-correction.}
}
\value{
if \code{plot = TRUE}, returns an object of class \code{KDE}, i.e. a
    list containing the following items:

\describe{
\item{x}{ horizontal plot coordinates}
\item{y}{ vertical plot coordinates}
\item{bw}{ the base bandwidth of the density estimate}
\item{ages}{ the data values from the input to the \code{kde} function}
}

or, if \code{x} has class \code{=detritals}, an object of class
\code{KDEs}, i.e. a list containing the following items:

\describe{
\item{kdes}{a named list with objects of class \code{KDE}}
\item{from}{the beginning of the common time scale}
\item{to}{the end of the common time scale}
\item{themax}{the maximum probability density of all the KDEs}
\item{xlabel}{the x-axis label to be used by \code{plot.KDEs}}
}
}
\description{
Creates one or more kernel density estimates using a combination of
the Botev (2010) bandwidth selector and the Abramson (1982)
adaptive kernel bandwidth modifier.
}
\examples{
data(examples)
kde(examples$DZ[['N1']],kernel="epanechnikov")
kde(examples$DZ,from=0,to=3000)
}
\references{
Abramson, I.S., 1982. On bandwidth variation in kernel estimates-a
square root law. The annals of Statistics, pp.1217-1223.

Botev, Z. I., J. F. Grotowski, and
D. P. Kroese. "Kernel density estimation via diffusion." The Annals
of Statistics 38.5 (2010): 2916-2957.

Vermeesch, P., 2012. On the visualisation of detrital age
distributions. Chemical Geology, 312, pp.190-194.
}
