% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{length.UPb}
\alias{length.UPb}
\alias{length.PbPb}
\alias{length.ArAr}
\alias{length.PD}
\alias{length.ThPb}
\alias{length.KCa}
\alias{length.ThU}
\alias{length.UThHe}
\alias{length.KDE}
\alias{length.KDEs}
\alias{length.fissiontracks}
\title{Count the number of aliquots in a sample}
\usage{
\method{length}{UPb}(x)

\method{length}{PbPb}(x)

\method{length}{ArAr}(x)

\method{length}{PD}(x)

\method{length}{ThPb}(x)

\method{length}{KCa}(x)

\method{length}{ThU}(x)

\method{length}{UThHe}(x)

\method{length}{KDE}(x)

\method{length}{KDEs}(x)

\method{length}{fissiontracks}(x)
}
\arguments{
\item{x}{an object of class \code{UPb}, \code{PbPb}, \code{ArAr},
\code{ThPb}, \code{KCa}, \code{PD}, \code{ThU},
\code{detritals}, \code{UThHe} or \code{fissiontracks}.}
}
\value{
a scalar
}
\description{
Count the number of rows (aliquots) in a dataset
    returned by \code{read.data} or, for datasets of class
    \code{detritals}, the number of columns (samples).
}
\examples{
data(examples)
length(examples$UPb)
}
