% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.data}
\alias{read.data}
\alias{read.data.default}
\alias{read.data.matrix}
\title{Read geochronology data}
\usage{
read.data(x, ...)

\method{read.data}{default}(x, method = "U-Pb", format = 1, ...)

\method{read.data}{matrix}(x, method = "U-Pb", format = 1, ...)
}
\arguments{
\item{x}{either a file name (\code{.csv} format) OR a matrix}

\item{...}{optional arguments to the \code{read.csv} function}

\item{method}{one of \code{'U-Pb'}, \code{'Ar-Ar'},
\code{'detritals'}, \code{Rb-Sr}, \code{Sm-Nd}, \code{Re-Os},
\code{'U-Th-He'}, \code{'fissiontracks'} or \code{'other'}}

\item{format}{formatting option, depends on the value of
    \code{method}.

if \code{method = 'Ar-Ar'}, then \code{format} is one of either:

\enumerate{
\item{\code{39/40, s[39/40], 36/40, s[36/40], 39/36, s[39/36]}}
\item{\code{39, 39/40, s[39/40], 36/40, s[36/40], 39/36, s[39/36]}}
}

if \code{method = 'fissiontracks'}, then \code{format} is one of
either:

\enumerate{
\item{the External Detector Method (EDM), which requires a
\eqn{\zeta}-calibration constant and its uncertainty, the induced
track density in a dosimeter glass, and a table with the
spontaneous and induced track densities.}

\item{LA-ICP-MS-based fission track data using the
\eqn{\zeta}-calibration method, which requires a 'session
\eqn{\zeta}' and its uncertainty and a table with the number of
spontaneous tracks, the area over which these were counted and one
or more U/Ca- or U-concentration measurements and their analytical
uncertainties.}

\item{LA-ICP-MS-based fission track data using the 'absolute
dating' method, which only requires a table with the the number of
spontaneous tracks, the area over which these were counted and one
or more U/Ca- or U-concentration measurements and their analytical
uncertainties.}
}}
}
\value{
an object of class \code{UPb}, \code{ArAr}, \code{UThHe},
    \code{ReOs}, \code{SmNd}, \code{RbSr}, \code{detritals},
    \code{fissiontracks} or \code{other}
}
\description{
Cast a \code{.csv} file or a matrix into one of \code{IsoplotR}'s
data classes
}
\details{
Example input files can be found by using \code{R}'s
\code{system.file(...)} function:

\enumerate{
\item \code{method = 'U-Pb'}:

\code{file.show(system.file("UPb.csv",package="IsoplotR"))}

\item \code{method = 'Ar-Ar'} and \code{format = 1}:

\code{file.show(system.file("ArAr1.csv",package="IsoplotR"))}

\item \code{method = 'Ar-Ar'} and \code{format = 2}:

\code{file.show(system.file("ArAr2.csv",package="IsoplotR"))}

\item \code{method = 'Re-Os'}:

\code{file.show(system.file("ReOs.csv",package="IsoplotR"))}

\item \code{method = 'Rb-Sr'}:

\code{file.show(system.file("RbSr.csv",package="IsoplotR"))}

\item \code{method = 'Sm-Nd'}:

\code{file.show(system.file("SmNd.csv",package="IsoplotR"))}

\item \code{method = 'U-Th-He'}:

\code{file.show(system.file("UThHe.csv",package="IsoplotR"))}

\item \code{method = 'fissiontracks'} and \code{format = 1}:

\code{file.show(system.file("FT1.csv",package="IsoplotR"))}

\item \code{method = 'fissiontracks'} and \code{format = 2}:

\code{file.show(system.file("FT2.csv",package="IsoplotR"))}

\item \code{method = 'fissiontracks'} and \code{format = 3}:

\code{file.show(system.file("FT3.csv",package="IsoplotR"))}

\item \code{method = 'detritals'}:

\code{file.show(system.file("DZ.csv",package="IsoplotR"))}
}
}
\examples{
# load one of the built-in .csv files:
data(examples)
concordia(examples$UPb)
}

