% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.matrix}
\alias{read.matrix}
\title{Read geochronology data}
\usage{
read.matrix(x, method = "U-Pb", format = 1)
}
\arguments{
\item{x}{a matrix}

\item{method}{see \code{read.data} for details}

\item{format}{see \code{read.data} for details}
}
\value{
see \code{read.data} for details
}
\description{
Cast a matrix into one of \code{IsoplotR}'s data classes
}
\examples{
# load one of the built-in .csv files:
fname <- system.file("UPb.csv",package="IsoplotR")
dat <- read.csv(fname,header=TRUE)
UPb <- read.matrix(dat,method='U-Pb',format=1)
concordia.plot(UPb)
}

