% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{UPb}
\alias{UPb}
\title{An example U-Pb dataset}
\description{
An example U-Pb dataset provided with Ludwig's Isoplot add-in
}
\details{
\code{UPb} is an object of class \code{UPb}, i.e. a list with two items

\code{x}: a matrix formatted according to \code{format}

\code{format}: an integer defining the format of \code{x}. Options are:

\code{1}: 7/6, s[7/6], 6/8, s[6/8], 7/5, s[7/5]
}
\examples{
data(UPb)
concordia.plot(UPb)
}
\author{
Ken Ludwig and Pieter Vermeesch
}

